requirements.txt
moderngl==5.8.2
PyQt5==5.15.9
numpy==1.24.3
Pillow==9.5.0